grid items layout fix
